--- .env.prod
@@ -1,1 +1,0 @@
-DRIVER=~fastapi
